+++ packages/react-ui-base/src/thread-dropdown/thread-item/thread-dropdown-thread-item.tsx
@@ -56,7 +56,7 @@ export const ThreadDropdownThreadItem = React.forwardRef<
     },
     props: mergeProps(componentProps, {
       type: "button",
-      "data-active": isActive || undefined,
+      "data-active": isActive ? "true" : undefined,
       onClick: handleClick,
     }),
   });